requirements.txt
discord.py
flask
PyNaCl